locking/percpu-rwsem: Fix kABI breakage
struct percpu_rw_semaphore breaks kABI as the semaphore is replaced by a
list of waiters. UEK_KABI_REPLACE macro was used to fix this to maintain
the correct struct size. When using debug config, we also include
dep_map in this as the whole created by rw_sem is bigger than dep_map
struct.

Orabug: 35732336
Signed-off-by: Harshvardhan Jha <harshvardhan.j.jha@oracle.com>
Signed-off-by: Lord Quas <quasimoto2000@gmail.com>